Ensure cgroup is deleted
