+++ cgroups/src/common.rs
@@ -3,6 +3,7 @@ use std::{
     fs::{self, File},
     io::{BufRead, BufReader, Write},
     path::{Path, PathBuf},
+    time::Duration,
 };
 
 use anyhow::{bail, Context, Result};
@@ -354,4 +355,22 @@ pub(crate) fn default_devices() -> Vec<LinuxDevice> {
             gid: None,
         },
     ]
+}
+
+pub(crate) fn delete_with_retry<P: AsRef<Path>>(path: P) -> Result<()> {
+    let mut attempts = 0;
+    let mut delay = Duration::from_millis(10);
+    let path = path.as_ref();
+
+    while attempts < 5 {
+        if fs::remove_dir(path).is_ok() {
+            return Ok(());
+        }
+
+        std::thread::sleep(delay);
+        attempts += attempts;
+        delay *= attempts;
+    }
+
+    bail!("could not delete {:?}", path)
 }
