+++ cgroups/src/v1/manager.rs
@@ -161,7 +161,7 @@ impl CgroupManager for Manager {
                     let _ = nix::sys::signal::kill(Pid::from_raw(pid), nix::sys::signal::SIGKILL);
                 }
 
-                util::delete_with_retry(cgroup_path.1)?;
+                common::delete_with_retry(cgroup_path.1)?;
             }
         }
